+++ client/it/src/it/java/net/awired/housecream/client/it/reserved/PinResourceReservedValueIT.java
@@ -15,14 +15,14 @@ public class PinResourceReservedValueIT {
 
     @Test
     public void should_not_get_value() throws Exception {
-        Integer pinValue = hcc.getPinResource().getValue(PIN_ID);
+        Float pinValue = hcc.getPinResource().getValue(PIN_ID);
 
         assertNull(pinValue);
     }
 
     @Test(expected = HccUpdateException.class)
     public void should_not_set_value() throws Exception {
-        hcc.getPinResource().setValue(PIN_ID, 1);
+        hcc.getPinResource().setValue(PIN_ID, 1f);
     }
 
 }
